refactor into a table component
