+++ Dashboard/src/Pages/OnCallDuty/SideMenu.tsx
@@ -6,8 +6,20 @@ import SideMenuItem from 'CommonUI/src/Components/SideMenu/SideMenuItem';
 import SideMenuSection from 'CommonUI/src/Components/SideMenu/SideMenuSection';
 import RouteMap, { RouteUtil } from '../../Utils/RouteMap';
 import PageMap from '../../Utils/PageMap';
+import Link from 'Common/Types/Link';
+import Navigation from 'CommonUI/src/Utils/Navigation';
 
 const DashboardSideMenu: FunctionComponent = (): ReactElement => {
+
+    let subItemMenuLink: Link | undefined = undefined;
+
+    if (Navigation.isOnThisPage(RouteMap[PageMap.ON_CALL_DUTY_EXECUTION_LOGS_TIMELINE]!)) {
+        subItemMenuLink = {
+            title: 'Timeline',
+            to: Navigation.getCurrentRoute(),
+        }
+    }
+
     return (
         <SideMenu>
             <SideMenuSection title="Overview">
@@ -32,6 +44,8 @@ const DashboardSideMenu: FunctionComponent = (): ReactElement => {
                         ),
                     }}
                     icon={IconProp.Logs}
+                    subItemIcon={IconProp.Clock}
+                    subItemLink={subItemMenuLink}
                 />
             </SideMenuSection>
         </SideMenu>
