+++ Dashboard/src/Utils/PageMap.ts
@@ -77,6 +77,7 @@ enum PageMap {
     ON_CALL_DUTY = 'ON_CALL_DUTY',
     ON_CALL_DUTY_POLICIES = 'ON_CALL_DUTY_POLICIES',
     ON_CALL_DUTY_EXECUTION_LOGS = 'ON_CALL_DUTY_EXECUTION_LOGS',
+    ON_CALL_DUTY_EXECUTION_LOGS_TIMELINE = 'ON_CALL_DUTY_EXECUTION_LOGS_TIMELINE',
     ON_CALL_DUTY_POLICY_VIEW = 'ON_CALL_DUTY_POLICY_VIEW',
     ON_CALL_DUTY_POLICY_VIEW_CUSTOM_FIELDS = 'ON_CALL_DUTY_POLICY_VIEW_CUSTOM_FIELDS',
     ON_CALL_DUTY_POLICY_VIEW_DELETE = 'ON_CALL_DUTY_POLICY_VIEW_DELETE',
